improve RSocketTestServer configuration
